CIRC-2266 refactoring
